Share button moved to FSH editor header
